Refactor the progress stuff
